测试SGIP协议，新增NodeId字段

+++ src/main/java/com/zx/sms/connect/manager/sgip/SgipEndpointEntity.java
@@ -10,7 +10,7 @@ public abstract class SgipEndpointEntity extends EndpointEntity {
 	private static final long serialVersionUID = -441048745116970563L;
 	private String loginName = GlobalConstance.emptyString;
 	private String loginPassowrd = GlobalConstance.emptyString;
-	private int nodeId = 0;
+	private long nodeId = 0;
 	public String getLoginName() {
 		return loginName;
 	}
@@ -23,10 +23,10 @@ public abstract class SgipEndpointEntity extends EndpointEntity {
 	public void setLoginPassowrd(String loginPassowrd) {
 		this.loginPassowrd = loginPassowrd;
 	}
-	public int getNodeId() {
+	public long getNodeId() {
 		return nodeId;
 	}
-	public void setNodeId(int nodeId) {
+	public void setNodeId(long nodeId) {
 		this.nodeId = nodeId;
 	}    
 }
